+++ inc/behaviac/base/string/stringutils.h
@@ -474,7 +474,7 @@ namespace behaviac
 
             if (end)
             {
-                int length = (end - params);
+                int length = (int)(end - params);
                 token = behaviac::string(params, length);
                 return end;
             }
@@ -496,7 +496,7 @@ namespace behaviac
 
                 if (end2)
                 {
-                    int length = (end2 - end);
+					int length = (int)(end2 - end);
                     token = behaviac::string(end, length);
                     return end2;
