+++ src/base/xml/xml.cpp
@@ -136,13 +136,13 @@ namespace behaviac
 	//! Get XML Node child nodes.
 	int CXmlNode::getAttrCount() const
 	{
-		return m_attributes.size();
+		return (int)m_attributes.size();
 	}
 
 	//! Get XML Node child nodes.
 	const char* CXmlNode::getAttr(int index) const
 	{
-		if ((uint32_t)index < m_attributes.size())
+		if ((size_t)index < m_attributes.size())
 		{
 			return m_attributes[index].GetValue();
 		}
@@ -684,7 +684,7 @@ namespace behaviac
 
 			behaviac::string out;
 			behaviac::StringUtils::Wide2Char(out, temp);
-			file->Write(out.c_str(), out.size());
+			file->Write(out.c_str(), (uint32_t)out.size());
 
 			return true;
 		}
